Move vitest.config.ts -> vitest.config.mts
