+++ examples/typescript-graphql-request/vitest.config.mts
@@ -1,12 +1,13 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../vitest.config.js';
+import { sharedConfig } from '../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'examples-persisted-documents-string-mode',
+      name: 'examples-typescript-graphql-request',
       include: ['**/*.spec.ts'],
+      testTimeout: 15_000,
     },
   }),
 );
