+++ packages/plugins/other/add/vitest.config.mts
@@ -1,13 +1,12 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../vitest.config.js';
+import { sharedConfig } from '../../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'examples-typescript-graphql-request',
+      name: 'add',
       include: ['**/*.spec.ts'],
-      testTimeout: 15_000,
     },
   }),
 );
