+++ packages/plugins/other/fragment-matcher/vitest.config.mts
@@ -1,5 +1,5 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../../../vitest.config.js';
+import { sharedConfig } from '../../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
