+++ packages/plugins/other/schema-ast/vitest.config.mts
@@ -1,11 +1,11 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../../../vitest.config.js';
+import { sharedConfig } from '../../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'typescript-resolvers',
+      name: 'schema-ast',
       include: ['**/*.spec.ts'],
     },
   }),
