+++ packages/plugins/typescript/operations/vitest.config.mts
@@ -1,11 +1,11 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../vitest.config.js';
+import { sharedConfig } from '../../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'examples-yoga-tests',
+      name: 'typescript-operations',
       include: ['**/*.spec.ts'],
     },
   }),
