+++ packages/plugins/typescript/typed-document-node/vitest.config.mts
@@ -1,11 +1,11 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../../../vitest.config.js';
+import { sharedConfig } from '../../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'time',
+      name: 'typed-document-node',
       include: ['**/*.spec.ts'],
     },
   }),
