+++ packages/presets/client/vitest.config.mts
@@ -1,12 +1,11 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../../vitest.config.js';
+import { sharedConfig } from '../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'graphql-modules-preset',
-      setupFiles: './vitest.setup.ts',
+      name: 'client-preset',
       include: ['**/*.spec.ts'],
     },
   }),
