+++ packages/presets/graphql-modules/vitest.config.mts
@@ -1,11 +1,12 @@
 import { defineProject, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../../../vitest.config.js';
+import { sharedConfig } from '../../../vitest.config.mjs';
 
 export default mergeConfig(
   sharedConfig,
   defineProject({
     test: {
-      name: 'introspection',
+      name: 'graphql-modules-preset',
+      setupFiles: './vitest.setup.ts',
       include: ['**/*.spec.ts'],
     },
   }),
