+++ tsconfig.spec.json
@@ -9,7 +9,7 @@
   },
   "exclude": [],
   "include": [
-    "**/vitest.config.ts",
+    "**/vitest.config.mts",
     "**/vitest.setup.ts",
     "**/*.spec.ts",
     "**/tests/**/*.ts",
